+++ src/app/mork/celery/celery_app.py
@@ -2,6 +2,7 @@
 
 import sentry_sdk
 from celery import Celery, signals
+from sentry_sdk.scrubber import DEFAULT_PII_DENYLIST, EventScrubber
 
 from mork import __version__
 from mork.conf import settings
@@ -20,6 +21,7 @@ app = Celery(
 def init_sentry(**_kwargs):
     """Initialize Sentry SDK on Celery startup."""
     if settings.SENTRY_DSN is not None:
+        pii_denylist = DEFAULT_PII_DENYLIST + ["email", "username"]
         sentry_sdk.init(
             dsn=settings.SENTRY_DSN,
             enable_tracing=True,
@@ -27,6 +29,8 @@ def init_sentry(**_kwargs):
             release=__version__,
             environment=settings.SENTRY_EXECUTION_ENVIRONMENT,
             max_breadcrumbs=50,
+            send_default_pii=False,
+            event_scrubber=EventScrubber(pii_denylist=pii_denylist),
         )
         sentry_sdk.set_tag("application", "celery")